# apps/backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/profiles": {
      "get": {
        "operationId": "ProfilesController_getProfiles",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List profiles",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProfileResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "profiles"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/profiles/{id}": {
      "put": {
        "operationId": "ProfilesController_updateProfile",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "profiles"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/profiles/{id}/role": {
      "put": {
        "operationId": "ProfilesController_updateUserRole",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Update user role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "profiles"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "AI Classroom Conversation Partner API",
    "description": "NestJS powered API with OpenAPI documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ProfileResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "User ID"
          },
          "createdAt": {
            "type": "string",
            "description": "Created at timestamp"
          },
          "updatedAt": {
            "type": "string",
            "description": "Updated at timestamp"
          },
          "fullName": {
            "type": "string",
            "description": "Full name"
          },
          "gender": {
            "type": "string",
            "description": "Gender"
          },
          "userRole": {
            "type": "string",
            "description": "User role",
            "enum": [
              "basic",
              "admin",
              "owner"
            ]
          },
          "conversationRole": {
            "type": "string",
            "description": "Conversation role"
          },
          "bio": {
            "type": "string",
            "description": "Bio"
          },
          "email": {
            "type": "string",
            "description": "Email address"
          },
          "confirmedAt": {
            "type": "string",
            "description": "Email confirmed at timestamp",
            "nullable": true
          }
        },
        "required": [
          "id",
          "createdAt",
          "updatedAt",
          "fullName",
          "gender",
          "userRole",
          "conversationRole",
          "bio",
          "email"
        ]
      },
      "UpdateProfileDto": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string",
            "nullable": true
          },
          "gender": {
            "type": "string",
            "nullable": true
          },
          "conversationRole": {
            "type": "string"
          },
          "bio": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "UpdateUserRoleDto": {
        "type": "object",
        "properties": {
          "userRole": {
            "type": "string",
            "enum": [
              "basic",
              "admin",
              "owner"
            ]
          }
        }
      },
...
    }
  }
}